// src/app/shared/shared/enum/contract/contract-period.enum.ts
/**
 * Enum các loại thời hạn hợp đồng
 * created by vhtruong - 07/08/2020
 */
export enum ContractPeriod {

    // 1 tháng
    OneMonth = 1,
    // 2 tháng
    TwoMonth = 2,
    // 3 tháng
    ThreeMonth = 3,
    // 6 tháng
    SixMonth = 4,
    // 1 năm
    OneYear = 5,
    // 6 năm
    ThreeYear = 6
}